Fixed invalid cast on mingw.
Fixes GCC error:
SkThread_win.cpp:39:61: error: invalid static_cast from type â€˜int32_t {aka int}â€™ to type â€˜const volatile LONG& {aka const volatile long int&}â€™

Review URL: https://codereview.chromium.org/17047004

git-svn-id: http://skia.googlecode.com/svn/trunk@9706 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/ports/SkThread_win.cpp
@@ -36,7 +36,7 @@ void sk_membar_aquire__after_atomic_dec() { }
 
 int32_t sk_atomic_conditional_inc(int32_t* addr) {
     while (true) {
-        LONG value = static_cast<LONG const volatile&>(*addr);
+        LONG value = static_cast<int32_t const volatile&>(*addr);
         if (value == 0) {
             return 0;
         }
